cast

+++ ucoin/src/log/ulog.c
@@ -83,7 +83,7 @@ void ulog_write(int Pri, const char* pFname, int Line, int Flag, const char *pTa
     va_list ap;
     time_t now = time(NULL);
     char tmstr[UCOIN_SZ_DTSTR + 1];
-    ucoin_util_strftime(tmstr, now);
+    ucoin_util_strftime(tmstr, (uint32_t)now);
 
     va_start(ap, pFmt);
     if (Flag) {
